Compute Approximate Information from Event Counts: Log Hazard Ratio

This function provides an asymptotic approximation to the information (i.e. precision, inverse of the variance) provided by a number of observed events pooled across treatment arms for a time-to-event outcome, analyzed using the log hazard ratio estimand. These functions may be useful in pre-trial planning to determine when analyses may occur under different assumptions about the nuisance parameters involved.

Usage

asymptotic_information_logrank(allocation_ratio = 1, total_events)

Arguments

allocation_ratio

A numeric scalar containing the allocation ratio of r participants to treatments for every 1 control. Defaults to 1.

total_events

A numeric vector containing the total number of events observed across both treatment arms.

Value

A numeric scalar or data.frame containing an approximate information level for the values of the inputs.

References

Schoenfeld, DA. 1983. "Sample-Size Formula for the Proportional-Hazards Regression Model." Biometrics 39 (2): 499. https://doi.org/10.2307/2531021.Mehta, CR, and Tsiatis AA. 2001. "Flexible Sample Size Considerations Using Information-Based Interim Monitoring". Drug Information Journal 35 (4): 1095–1112. https://doi.org/10.1177/009286150103500407

See also

Examples

asymptotic_information_logrank(
  allocation_ratio = 1,
  total_events = 90
)
#> [1] 22.5

asymptotic_information_logrank(
  allocation_ratio = 1,
  total_events = c(66, 90)
)
#>   allocation_ratio total_events information_asymptotic
#> 1                1           66                   16.5
#> 2                1           90                   22.5

asymptotic_information_logrank(
  allocation_ratio = c(1, 2),
  total_events = c(66, 90)
)
#>   allocation_ratio total_events information_asymptotic
#> 1                1           66               16.50000
#> 2                2           66               14.66667
#> 3                1           90               22.50000
#> 4                2           90               20.00000