Parameters for using boot::boot() in Non-Sequential Analyses

These functions are convenience functions for supplying parameters to bootstrap-based analyses. NOTE: sequential analyses require computing the covariance of estimates using a bootstrap function that preserves the covariance inherent in group sequential designs. This function should not be used for group sequential designs.

Usage

boot_control(
  bootstrap_n = 10000,
  parallel = "no",
  ncpus = getOption("boot.ncpus", 1L),
  ...
)

boot_control_testing(bootstrap_n = 1000, ...)

Arguments

bootstrap_n

A numeric scalar indicating the number of bootstrap replicates to perform.

parallel

A character scalar indicating whether and how to use parallel computing: see ?boot

ncpus

A numeric scalar indicating the number of cores to use in parallel computing: see ?boot

...

Other named parameters passed to boot

Value

A list containing the default parameters or supplied alternatives.

See also

Examples

boot_control()
#> $R
#> [1] 10000
#> 
#> $parallel
#> [1] "no"
#> 
#> $ncpus
#> [1] 1
#> 
boot_control_testing()
#> $R
#> [1] 1000
#> 
#> $parallel
#> [1] "no"
#> 
#> $ncpus
#> [1] 1
#>