colon_cancer: Processed Moerton Colon Cancer Data

Data from an adjuvant chemotherapy trial for colon cancer with variables renamed and labeled, missing covariates imputed, and time scale converted to years. For more information, see ?survival::colon.

Usage

colon_cancer

Format

colon_cancer

A data frame with 929 rows and 15 columns:

.id

Participant ID

arm

Treatment Assignment

age

Patient age

sex

Patient sex

obstruction

Colon obstruction by tumor

perforation

Colon perforation by tumor

organ_adherence

Organ adherence by tumor

positive_nodes

Lymph nodes with detectable cancer

differentiation

Differentiation of tumor

local_spread

Extent of local spread of tumor

time_surgery_registration

Time from surgery to registration

event_death

Death on study (1 = yes, 0 = no)

event_recurrence

Recurrence (1 = yes, 0 = no)

years_to_death

Time to death/censoring in years

years_to_recurrence

Time to recurrence/censoring in years