Count outcome events from prepared study data

This function uses the timing of events, such as randomization times and times of outcomes, to reconstruct the data available at an earlier point in the study. This can be used for monitoring the rate of information accrual over the course of a study. This function returns the count of outcomes by study time.

Usage

count_outcomes(prepared_data, study_time = NULL)

Arguments

prepared_data

A prepeared dataset: see prepare_monitored_study_data

study_time

A numeric scalar indicating the study time at which the data should be reconstructed and events should be counted

Value

A data.frame in long format, with one row per individual per event (i.e. randomization, outcomes).

Examples

# To be added