Estimate the observed information level

This is a function for computing the information level from a dataset. For multi-stage designs, earlier results can be passed as a monitored_design object.

Usage

estimate_information(
  data,
  monitored_design,
  estimation_function,
  estimation_arguments,
  correction_function = NULL,
  orthogonalize = NULL,
  rng_seed,
  return_results = FALSE,
  control = monitored_analysis_control()
)

Arguments

data

A data.frame containing the data to be analyzed. A column named .id indicates which observations correspond to each individual.

monitored_design

An object of class monitored_design created using initialize_monitored_design()

estimation_function

A function whose arguments include a data.frame named data

estimation_arguments

A list of any additional arguments needed by estimation_function

correction_function

A function which takes named arguments from estimation_arguments and returns a numeric scalar: this is used to scale the variance before computing information.

orthogonalize

Logical scalar: Should estimates, their covariance, and the resulting test statistics be orthogonalized?

rng_seed

Numeric scalar containing the L'Ecuyer pseudorandom number generator seed

return_results

Logical scalar: Should estimates be returned, or only the covariance and information levels?

control

A list containing the control arguments for computation, typically created with monitored_analysis_control()

Value

A list containing the following elements:

See also

Examples

# To be added