Adjust Information Target for a Group Sequential Design

Group Sequential Design methods can control the Familywise Type I Error (FWER) due to multiple assessments of efficacy and allow for futility stopping. Interim stopping can potentially reduce the average length of a study, often requiring only a modest increase in the maximum potential sample size. The method used to adjust the sample size of a single stage design to account for multiplicity can also be used to adjust the information target of an information monitored design.

Usage

required_information_sequential(information_single_stage, trial_design)

get_gsd_inflation_factor(trial_design)

Arguments

information_single_stage

A numeric scalar containing the information level for an information monitored design with a single efficacy analysis, created using required_information_single_stage

trial_design

An object of type trialDesignGroupSequential created by rpact::getDesignGroupSequential containing a group sequential design specification

Value

get_gsd_inflation_factor returns a numeric scalar containing the inflation factor, and required_information_sequential returns a numeric scalar containing the information target accounting for the pre-planned analyses specified in trial_design.

Details

It is often advantageous to perform pre-planned interim analyses during a trial to assess whether the efficacy of treatments has already been established, or whether it is futile to continue collecting data. Group Sequential Designs are a principled way of carrying out such studies that control familywise type I error. The use of \(\alpha\)-spending functions allow for stopping boundaries to be adjusted to the amount of data collected.

When there are assessments for efficacy or futility, achieving the same level of power and type I error requires an increase in sample size. This "inflation factor" depends on the number of analyses (often denoted \(K\)) and type of stopping boundaries chosen (often denoted as a "shape parameter" \(\Delta\)).

get_gsd_inflation_factor extracts the "inflation factor" from a group sequential design created by rpact::getDesignGroupSequential. required_information_sequential takes in an information target \(\mathcal{I}\) for a single-stage design (no interim analyses of efficacy), and a group sequential design specification, and returns an information target \(\mathcal{I}_{GSD}\) for an information monitored design with group sequential stopping for efficacy and/or futility.

References

Mehta, CR, and Tsiatis AA. 2001. "Flexible Sample Size Considerations Using Information-Based Interim Monitoring". Drug Information Journal 35 (4): 1095–1112. https://doi.org/10.1177/009286150103500407Jennison, C, and Turnbull, BW. 1999. Group Sequential Methods with Applications to Clinical Trials. Chapman; Hall/CRC. https://doi.org/10.1201/9780367805326.Wassmer, G, and Pahlke, F. 2025. Rpact: Confirmatory Adaptive Clinical Trial Design and Analysis. https://doi.org/10.32614/CRAN.package.rpact.

See also

Examples

# To be added