Initialize an Information Monitoring Design

Rather than planning data collection until a pre-specified sample size is reached based on estimates of nuisance parameters, an information-monitored study continues data collection until they provide sufficient precision to identify a meaningful difference with appropriate power and control of Type I Error. The design is specified in terms of an information level rather than a fixed sample size, and recruitment is stopped when the accrued participants are projected to meet the required level of precision to test the hypotheses of interest. This function creates an object that encapsulates all aspects of the study design that must be specified in advance, such as the sequential analysis plan, the value of the estimand under the null, the maximum sample size allowed to be enrolled before recruitment is automatically suspended, the target information level, whether test statistics should be orthogonalized, and the pseudorandom number generator seeds to be used in analyses.

Usage

initialize_monitored_design(
  trial_design,
  null_value = NULL,
  maximum_sample_size,
  information_target,
  orthogonalize,
  rng_seed_analysis
)

Arguments

trial_design

An object of type trialDesignGroupSequential created by rpact::getDesignGroupSequential() containing a group sequential design specification

null_value

Numeric scalar containing the value of the estimand under the null hypothesis, e.g. 0 for differences or 1 for ratios

maximum_sample_size

Numeric scalar containing the maximum sample size after which recruitment will be stopped if it has not already been stopped from information monitoring.

information_target

A numeric scalar containing the information level for an information monitored design created using required_information_single_stage() and adjusted for multiplicity using required_information_sequential()

orthogonalize

Logical scalar: Should estimates, their covariance, and the resulting test statistics be orthogonalized?

rng_seed_analysis

Numeric scalar containing the L'Ecuyer pseudorandom number generator seed to be used for the analyses specified in trial_design

Value

A "monitored design" object which is used in several impart functions.

Examples

# To be added