Control arguments for performing information monitored analyses

This provides a way to set and modify the analysis parameters for information monitoring.

Usage

monitored_analysis_control(
  orthogonal_and_oblique = FALSE,
  required_packages = character(0),
  n_bootstrap = 1000,
  n_cores = 1,
  use_load_balancing = FALSE
)

monitored_analysis_control_testing(
  orthogonal_and_oblique = FALSE,
  required_packages = character(0),
  n_bootstrap = 250,
  n_cores = 1,
  use_load_balancing = FALSE
)

Arguments

orthogonal_and_oblique

A logical scalar: if orthogonal == TRUE should non-orthogonalized (oblique) test statistics and decisions be returned?

required_packages

A character vector of required packages to be loaded on a cluster: only used when n_cores > 1.

n_bootstrap

Scalar number containing the number of bootstrap replicates to perform

n_cores

Scalar number of cores to use.

use_load_balancing

Logical scalar: Should load balancing be used?

Value

A list containing the default parameters or supplied alternatives.

See also

Examples

# To be added