Approximate information from a random samples of a given size for log hazard ratio

Approximate information from a random samples of a given size for log hazard ratio

Usage

asymptotic_information_logrank(allocation_ratio = 1, total_events)

Arguments

allocation_ratio

Numeric scalar containing the allocation ratio of r treatments to 1 control. Defaults to 1.

total_events

Numeric vector containing the total number of events observed across both treatment arms

Value

A numeric scalar or data.frame containing an approximate information level for the values of the inputs.

Examples

asymptotic_information_logrank(
  allocation_ratio = 1,
  total_events = 90
)
#> [1] 22.5

asymptotic_information_logrank(
  allocation_ratio = 1,
  total_events = c(66, 90)
)
#>   allocation_ratio total_events information_asymptotic
#> 1                1           66                   16.5
#> 2                1           90                   22.5

asymptotic_information_logrank(
  allocation_ratio = c(1, 2),
  total_events = c(66, 90)
)
#>   allocation_ratio total_events information_asymptotic
#> 1                1           66               16.50000
#> 2                2           66               14.66667
#> 3                1           90               22.50000
#> 4                2           90               20.00000