Compute stanadardization (i.e. G-Computation) estimator

Computes the estimate of a contrast of means for continuous and binary outcomes on the additive scale, relative scale. Computing a marginal odds ratio is also possible for binary outcomes.

Usage

standardization(
  data,
  estimand = "difference",
  y0_formula,
  y1_formula,
  family,
  treatment_column = NULL,
  outcome_indicator_column = NULL
)

standardization_correction(
  data,
  y0_formula,
  y1_formula,
  treatment_column,
  outcome_indicator_column
)

Arguments

data

A data.frame containing baseline covariates (e.g. x1, x2, ...), a binary treatment indicator (e.g. tx where 1 = Treatment; 0 = Control), outcome variables (e.g. y1, y2, ...), and outcome indicators (e.g. .r_1, .r_2, ...). The outcome indicators indicate whether an outcome has been observed (1), is missing (0), or not yet obtained (NA).

estimand

A character scalar: "difference" (for a difference in means or risk difference), "ratio" (for a ratio of means or relative risk), or "oddsratio" (for an odds ratio for a binary outcome).

y0_formula

A stats::formula specifying the relationship between the outcome and covariates in the control arm.

y1_formula

A stats::formula specifying the relationship between the outcome and covariates in the treatment arm.

family

The stats::family for the outcome regression model

treatment_column

A character scalar indicating the column containing the treatment indicator.

outcome_indicator_column

A character scalar indicating the column containing the outcome indicator corresponding to treatment_column.

Value

A list containing the marginal means and their contrast:

Details

Generalized linear models, stratified by treatment arm, are fitted using the specified formulas. Predictions are generated for each individual from each treatment model, representing the predicted outcome under each treatment assignment. These predictions are marginalized over the covariates by averaging to produce marginal estimates of the means. Finally, the contrast is computed and returned.

Variance estimates can be obtained using the nonparametric bootstrap. Unadjusted estimates can be obtained by using intercept only models.

Examples

ex_1 <- example_1
ex_1$.r_4 <- 1*(!is.na(ex_1$y_4))

standardization(
  data = ex_1,
  estimand = "difference",
  y0_formula = y_4 ~ x_1 + x_2 + x_3 + x_4,
  y1_formula = y_4 ~ x_1 + x_2 + x_3 + x_4,
  family = gaussian,
  treatment_column = "tx",
  outcome_indicator_column = ".r_4"
)
#> $estimate
#> [1] 6.296167
#> 
#> $y1_pred
#>          1          2          3          4          5          6          7 
#> -0.2239641  7.0957436 12.1894446 -0.8718126  9.2167704  1.9733213  4.0436358 
#>          8          9         10         11         12         13         14 
#>  5.6965007 -2.4992619 13.2944272  3.7067169 11.4639367 11.9251845  0.9520889 
#>         15         16         17         18         19         20         21 
#>  5.8057447 10.2435111 10.1713020 14.7419475 -3.1686397  9.8851426 10.6235739 
#>         22         23         24         25         26         27         28 
#> 15.0371821 10.6876981 11.3359684  9.5660750  6.1220109  1.0872081 16.8303605 
#>         29         30         31         32         33         34         35 
#> -5.2155383 -3.6412579  1.6599420  0.2836910  7.8771309 -2.4109066 13.6579311 
#>         36         37         38         39         40         41         42 
#>  1.4909818  3.2010649 12.7601123  4.3743874  7.0817270  4.0015711  6.9432138 
#>         43         44         45         46         47         48         49 
#> 22.3205928 -1.5927481  6.2890834 12.6578964  3.2265957  1.6721800  2.7993724 
#>         50         51         52         53         54         55         56 
#> 10.3832054 -0.5178242  8.9596713 10.1469373  2.7553902  5.2257128 -0.5645766 
#>         57         58         59         60         61         62         63 
#>  6.7555997 21.3417070 12.1320190 -2.3065519 17.8513514 11.7879040 -0.9753888 
#>         64         65         66         67         68         69         70 
#> -7.6898240 11.7287195  7.5419467  3.1458171  5.7445784 12.9490940  7.2072794 
#>         71         72         73         74         75         76         77 
#>  4.2832024  1.5941017  4.8687947  8.5841941  8.7926314  0.8479647 12.3342891 
#>         78         79         80         81         82         83         84 
#> -4.3988793  2.2017547 15.6548252  8.6561544  1.9228489  6.1680674 15.2493156 
#>         85         86         87         88         89         90         91 
#> 11.8443673  5.3836921  5.3080378  1.4140813  1.5321803  4.0765965  7.5932509 
#>         92         93         94         95         96         97         98 
#> 12.0331944 -6.3222710  2.9801292 18.0965850  0.7594923  0.4559849  6.8596737 
#>         99        100        101        102        103        104        105 
#>  1.2234896 14.2204941  4.5737059 -0.1009448  5.8956232  7.1283963  4.4465268 
#>        106        107        108        109        110        111        112 
#>  0.8400922  6.3977719 -5.4111876 -1.3457735  3.3757373 13.7703833  0.7138871 
#>        113        114        115        116        117        118        119 
#>  4.3584600  6.0534926  3.1938314 14.9168198  3.6270263  5.4441931 -1.7688350 
#>        120        121        122        123        124        125        126 
#>  9.3303201 -0.3428776  8.9739171  5.9257101  8.7455360  4.3738201  3.8895499 
#>        127        128        129        130        131        132        133 
#>  5.0173971 -9.7863053 -0.2572515 10.5191010 17.9343578  5.4614891  0.5901794 
#>        134        135        136        137        138        139        140 
#> -6.1977687 12.4264452  0.6181823  5.9381079  2.0401745  7.0324661  4.1015194 
#>        141        142        143        144        145        146        147 
#> -3.8481362 12.9019927  5.1554420 11.2359320  5.7402619 19.1036244 -9.0173167 
#>        148        149        150        151        152        153        154 
#>  3.1769958  8.0773865  4.4442425  8.3334870  0.3701447  9.6476652 15.7871523 
#>        155        156        157        158        159        160        161 
#>  7.7055956 -2.0592089  6.2930691 -0.9380754 -6.1496674  5.7179982 15.7850848 
#>        162        163        164        165        166        167        168 
#> 12.5817826  1.8793989  6.3092396  7.9384870 15.7822948 11.7664872  6.0069891 
#>        169        170        171        172        173        174        175 
#> 19.1931515 17.6652281 -2.1087632  1.8039543 11.0874957 10.8103864  2.1043230 
#>        176        177        178        179        180        181        182 
#> 19.7635214  3.0559390 13.9169616  8.8170221 13.0485840 20.6582016 12.0393958 
#>        183        184        185        186        187        188        189 
#>  9.6322568  1.8881862  7.8234821  1.7308217 12.1737809  1.8274418  6.0756956 
#>        190        191        192        193        194        195        196 
#> -2.6286500  8.0257412  6.6674861  8.8607391  6.8201389  9.6698391  3.6753270 
#>        197        198        199        200        201        202        203 
#> 14.3635332 -4.6998419  7.2932344  3.6080537 15.1565731 18.3685325  6.2330818 
#>        204        205        206        207        208        209        210 
#> -0.8077598 11.1871590 -1.3061570  5.0274634 13.9005531  9.8602956  4.1956657 
#>        211        212        213        214        215        216        217 
#>  4.1762135  7.1487749 10.1210473  6.1556535 18.0910439  3.8775112 13.2430854 
#>        218        219        220        221        222        223        224 
#> -6.7837228 10.0432117 -0.4403344  3.0502343 20.2128429 12.8000042 15.2499824 
#>        225        226        227        228        229        230        231 
#> 10.0925128 18.3300147 -3.2759231  1.2296909 13.7889406 -7.5119440 13.9784653 
#>        232        233        234        235        236        237        238 
#>  3.4534868  9.1552946 14.4305496  0.3009887 13.5937682  7.1104274 -3.4429610 
#>        239        240        241        242        243        244        245 
#> 10.0351136 12.3825489 14.5615649 22.2031975  8.1265358 -4.7599712  9.1004180 
#>        246        247        248        249        250        251        252 
#>  8.5111136  8.1923861 10.4746035 -0.6923349  0.8163619  8.4835868 -1.9226080 
#>        253        254        255        256        257        258        259 
#> 10.4299795 12.0549192  1.6047981  7.1785874 11.8072717  4.8068547  3.6308867 
#>        260        261        262        263        264        265        266 
#>  8.2975850  3.7725881  4.9119119  5.3530171 10.1811346 13.9582373 12.0347794 
#>        267        268        269        270        271        272        273 
#> 14.4376021  7.6041340 10.3805664  8.3708947 -5.4165687 15.4286886  3.5774313 
#>        274        275        276 
#>  5.5025500  8.4439165  0.8434968 
#> 
#> $y0_pred
#>           1           2           3           4           5           6 
#> -3.20959348  1.67173358 -0.11351757  0.15455387 -0.67578923 -3.62646575 
#>           7           8           9          10          11          12 
#>  1.03316505  1.58774554 -6.61096489  6.55832412 -0.07863326  3.18346092 
#>          13          14          15          16          17          18 
#>  4.57195321 -1.56130432  0.57767645  1.50051504  3.61232178  2.44493472 
#>          19          20          21          22          23          24 
#> -0.58993140  1.26185458  3.03427072  6.95298710  0.97662200 -1.28664520 
#>          25          26          27          28          29          30 
#>  2.50207287 -0.84740752 -0.91305418  2.94617737 -4.42485254 -6.16924774 
#>          31          32          33          34          35          36 
#> -0.27794384 -0.36802468 -1.80522335 -4.56094475  2.04709778 -0.76864337 
#>          37          38          39          40          41          42 
#>  0.71752253  1.56973575 -2.86683866  0.75768370  2.84154217 -0.07591796 
#>          43          44          45          46          47          48 
#>  8.15336341 -2.51631379  0.11250264  2.11769843 -0.33670127 -5.06433524 
#>          49          50          51          52          53          54 
#> -0.91070696  1.48910201 -0.74654849  1.85770312 -1.12744705 -3.66332489 
#>          55          56          57          58          59          60 
#>  0.44269740 -2.85442621 -1.21711472  8.39183054  3.62008936 -2.34811478 
#>          61          62          63          64          65          66 
#>  5.70433161  4.93228190 -4.30689607 -4.97072571  4.02192956 -2.04698333 
#>          67          68          69          70          71          72 
#> -0.10311566  0.01238429  1.52562273  2.74724573  0.03617009 -0.18001034 
#>          73          74          75          76          77          78 
#> -1.28809244  2.89856411 -1.67701163  2.33115960  5.25812845 -4.50550015 
#>          79          80          81          82          83          84 
#> -0.69238009  2.02195467  0.64428702 -2.14432204  0.75331337  3.10633058 
#>          85          86          87          88          89          90 
#>  4.38959454 -2.69163460 -2.43411496 -1.24312530 -2.07231539  0.46625221 
#>          91          92          93          94          95          96 
#>  2.53071997  0.41872003 -3.37688378 -0.33926467  2.19416724  0.07617651 
#>          97          98          99         100         101         102 
#> -3.78015351 -0.68491791 -3.36321878  4.16538293  2.51074816 -2.38489973 
#>         103         104         105         106         107         108 
#> -1.83145110 -0.26850932 -2.16967886 -4.13407076 -0.65625803 -3.93130000 
#>         109         110         111         112         113         114 
#> -2.70375274 -3.29560553  2.56847051 -1.58686839 -2.06212160  0.07283831 
#>         115         116         117         118         119         120 
#> -2.03957464  3.05492224 -2.71382952 -1.25133143 -3.81534965 -2.54936522 
#>         121         122         123         124         125         126 
#> -2.38426726  2.64947577  0.95738720  1.70171380 -1.61346935  1.02107931 
#>         127         128         129         130         131         132 
#> -6.62187491 -6.02643801 -3.62371761  0.58999635  7.16654705 -0.55612712 
#>         133         134         135         136         137         138 
#> -1.88812412 -6.14827877  3.03053104 -4.23220473  0.52878607 -4.13751643 
#>         139         140         141         142         143         144 
#> -1.03422480 -0.19575460 -4.52532883  2.47060740  2.76087541  5.16086733 
#>         145         146         147         148         149         150 
#>  1.94920519  4.32579204 -4.08388662 -3.45716282  2.24413483 -0.04342910 
#>         151         152         153         154         155         156 
#>  0.41189559 -1.43490413  3.52719436  1.37038952  2.24786546  0.73732831 
#>         157         158         159         160         161         162 
#>  0.69877800 -2.09199078 -1.17518710  2.58596673  5.19337269  2.84895369 
#>         163         164         165         166         167         168 
#> -4.14472469  0.20620912 -0.20541768  5.55647905  2.75204335 -1.11877024 
#>         169         170         171         172         173         174 
#>  7.47981508  4.89695769 -4.68623965 -1.42553567  1.61767118  1.73503948 
#>         175         176         177         178         179         180 
#> -4.74246299  6.06190883  2.02325901  5.42966442  5.06193113  3.43850348 
#>         181         182         183         184         185         186 
#>  8.93386090  0.89177121  1.36180437 -4.20139804 -1.42896777 -1.11815397 
#>         187         188         189         190         191         192 
#>  2.99197362  0.20901270  0.09375036 -2.05973249 -1.27459595 -2.69899303 
#>         193         194         195         196         197         198 
#>  4.22002503  1.18250121  2.34576694 -0.74395391  5.81776551 -4.77487268 
#>         199         200         201         202         203         204 
#> -1.46002401  0.91555868  3.92249991  5.54096512  0.75842524 -2.72562664 
#>         205         206         207         208         209         210 
#>  2.98022493 -3.84030039 -2.64938167  3.10667299 -2.34632578 -1.47239333 
#>         211         212         213         214         215         216 
#> -1.40409206  1.44806099  4.24654287 -0.52302664  7.11246048 -0.71879738 
#>         217         218         219         220         221         222 
#>  0.02352463 -5.81168379  1.42977951 -4.37622588 -0.70221159  4.13780833 
#>         223         224         225         226         227         228 
#>  4.31163099  5.44577413  1.84815656  5.30412746 -2.88941800 -2.18148510 
#>         229         230         231         232         233         234 
#> -0.52203620 -4.92770919  2.91609057 -1.53631440  1.43042865  2.05635781 
#>         235         236         237         238         239         240 
#> -2.53719903  1.11197940  2.87031591 -3.11790988  2.10022269  1.42066250 
#>         241         242         243         244         245         246 
#>  4.11929851  3.74568861 -3.45294570 -3.26846245  0.81281859  2.16753277 
#>         247         248         249         250         251         252 
#>  2.29829620 -1.85372898 -0.29727354 -3.95772248  4.81244768 -2.39220378 
#>         253         254         255         256         257         258 
#>  0.86048587  4.72523864 -1.18458136  1.46750255  1.19636642  0.47106236 
#>         259         260         261         262         263         264 
#> -3.23201323  1.47602694 -4.59470853  0.10330185  0.89203139  1.17129894 
#>         265         266         267         268         269         270 
#>  1.95575044  0.88808492  1.41799273  2.43133959 -1.92088034  3.36695935 
#>         271         272         273         274         275         276 
#> -5.46305388  4.79872522 -0.98257703  0.54406939  0.27245642 -2.91536021 
#> 
#> attr(,"class")
#> [1] "standardization"